+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+* Fixed unreported output files from the same source asset
 * Fixed *ShaderCompilerData* parsing in 2021.2.0a16 or newer
 
 ## [0.7.0-preview] - 2021-11-29